requirements.txt
ipykernel
langchain
langchain_community
streamlit
python-dotenv
langchain-ollama
langchain-groq
faiss-cpu
pypdf
